% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R, R/intsp-class.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Conversion from intgrd to data frame.}
\arguments{
\item{x}{an object of class intgrd}

\item{x}{an object of class intsp}
}
\value{
an object of class data.frame

an object of class data.frame
}
\description{
This function simply converts an intgrd object to a SpatialPixelsDataFrame
object prior to calling as.data.frame.SpatialPixelsDataFrame.

This function simply converts an intsp object to a SpatialPointsDataFrame
object prior to calling as.data.frame.SpatialPointsDataFrame.
}
