% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\docType{methods}
\name{$<-}
\alias{$<-}
\alias{$<-,intgrd-method}
\alias{$<-,intsp-method}
\title{Atomic assignment of interval spatial objects.}
\usage{
\S4method{$}{intgrd}(x, name) <- value

\S4method{$}{intsp}(x, name) <- value
}
\arguments{
\item{x}{An object of class \code{intsp} or class \code{intgrd}.}

\item{name}{The variable name on which to subset. This variable name can be
located in the data, coordinates, or interval slot.}

\item{value}{The new data used to replace the existing data in the desired slot.}
}
\description{
Atomic assignment of interval spatial objects.
}
