% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{$<-}
\alias{$<-}
\title{Atomic assignment operator for intsp data}
\arguments{
\item{x}{an object of class intsp}

\item{name}{the variable name from which to subset}

\item{value}{values to replace the current data in the variable}
}
\value{
a vector containing the requested data.
}
\description{
This method extends the traditional \'$\' operator by also looking in the
interval slot for the requested variable.
}
