% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,intgrd-method}
\alias{$,intsp-method}
\title{Atomic sub-setting of interval spatial objects.}
\usage{
\S4method{$}{intgrd}(x, name)

\S4method{$}{intsp}(x, name)
}
\arguments{
\item{x}{An object of class \code{intsp} or class \code{intgrd}.}

\item{name}{The variable name on which to subset. This variable name can be
located in the data, coordinates, or interval slot.}
}
\value{
The requested information from the data, coordinate,
  or interval slot.
}
\description{
Atomic sub-setting of interval spatial objects.
}
