% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat-extensions.R
\name{fit.intvariogram}
\alias{fit.intvariogram}
\title{Function to create a variogram object for interval-valued data.}
\usage{
fit.intvariogram(x, models = gstat::vgm(rep("Sph", 3)), ...)
}
\arguments{
\item{x}{An object of class \code{intvariogram}.}

\item{models}{an object of class \code{variogramModelList}. The user must
specify at least two variogram models to fit (for center and radius).
If less than three models are specified then the method does not fit a
variogram for the center/radius interaction.}

\item{...}{Additional arguments for \code{\link[gstat]{fit.variogram}}.}
}
\value{
A list of variograms objects from the gstat package.
}
\description{
Function to create a variogram object for interval-valued data.
}
