% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{head.intgrd}
\alias{head.intgrd}
\title{Print the head of an intgrd object.}
\usage{
\method{head}{intgrd}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{An object of class \code{intgrd}.}

\item{n}{Number of rows to print to the screen.}

\item{...}{Additional arguments to \code{print}.}

\item{digits}{Determines how numeric values are printed to the screen
(default from \code{sp} package).}
}
\value{
Prints a subset of the object observations to the screen,
  identical to \code{\link[sp]{SpatialPoints-class}}, as well as
  summary statistics for the interval slot.
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
