% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{interval<-}
\alias{interval<-}
\alias{interval<-,intgrd-method}
\alias{interval<-,SpatialPixelsDataFrame-method}
\alias{interval<-,intsp-method}
\alias{interval<-,SpatialPointsDataFrame-method}
\title{Function to reassign the contents of the interval slot}
\usage{
interval(x) <- value

\S4method{interval}{intgrd}(x) <- value

\S4method{interval}{SpatialPixelsDataFrame}(x) <- value

\S4method{interval}{intsp}(x) <- value

\S4method{interval}{SpatialPointsDataFrame}(x) <- value
}
\arguments{
\item{x}{An object of class intsp or intgrd.}

\item{value}{Either a character vector of length two specifying
the column names which will occupy the interval slot. Or, a matrix
of two columns to fill the slot.}
}
\description{
Function to reassign the contents of the interval slot
}
