% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{interval}
\alias{interval}
\alias{interval,intgrd-method}
\alias{interval,intsp-method}
\title{Function to extract the interval of an intsp or intgrd object}
\usage{
interval(x)

\S4method{interval}{intgrd}(x)

\S4method{interval}{intsp}(x)
}
\arguments{
\item{x}{An object of class intsp or intgrd.}
}
\value{
A matrix containing the interval data.
}
\description{
Function to extract the interval of an intsp or intgrd object
}
