% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intsp-class.R
\name{interval<-}
\alias{interval<-}
\title{Function to set the value of the interval slot in an intsp or intgrd object}
\arguments{
\item{x}{an object of class intsp or intgrd}

\item{value}{either column names specifying the location of the interval
data or a matrix of two columns with the interval data}

\item{x}{an object of class intsp}

\item{x}{an object of class SpatialPointsDataFrame}
}
\value{
an object of class intsp with a reassigned interval slot

an object of class intsp
}
\description{
Function to set the value of the interval slot in an intsp or intgrd object

Reassign values in the interval slot

Convert a SpatialPointsDataFrame object to an intsp object.
}
