% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intsp-class.R
\name{interval}
\alias{interval}
\title{Function to extract the interval of an intsp or intgrd object}
\arguments{
\item{x}{an object of class intsp or intgrd}

\item{x}{an object of class intsp}
}
\value{
the interval slot of an intsp object
}
\description{
Function to extract the interval of an intsp or intgrd object

Create an intsp class object
}
