% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{class}
\name{intgrd-class}
\alias{intgrd-class}
\alias{intgrd}
\title{An interval extension of a SpatialPixelsDataFrame}
\description{
An interval extension of a SpatialPixelsDataFrame
}
\section{Slots}{

\describe{
\item{\code{interval}}{A matrix of two columns representing the lower and upper
endpoints of an interval.}
}}

