% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{class}
\name{intsp-class}
\alias{intsp-class}
\alias{intsp}
\title{An interval extension of a SpatialPointsDataFrame}
\description{
An interval extension of a SpatialPointsDataFrame
}
\section{Slots}{

\describe{
\item{\code{interval}}{A matrix of two columns representing the lower and upper
endpoints of an interval.}
}}

