% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{intvariogram}
\alias{intvariogram}
\alias{intvariogram,intgrd-method}
\alias{intvariogram,intsp-method}
\title{Function to fit empirical variograms for an interval-valued spatial object}
\usage{
intvariogram(x, centerFormula = center ~ 1, ...)

\S4method{intvariogram}{intgrd}(x, centerFormula = center ~ 1, ...)

\S4method{intvariogram}{intsp}(x, centerFormula = center ~ 1, ...)
}
\arguments{
\item{x}{An object of class intsp or intgrd.}

\item{centerFormula}{A formula specifying any external trends for the
interval centers.}

\item{...}{Additional arguments for sp::variogram().}
}
\value{
An object of class 'intvariogram' containing empirical variograms
  for the center, radius, and center/radius interaction.
}
\description{
Function to fit empirical variograms for an interval-valued spatial object
}
