% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{intvariogram}
\alias{intvariogram}
\title{Function to fit empirical variograms for an interval-valued spatial object}
\arguments{
\item{x}{an object of class intsp or intgrd}

\item{x}{an object of class intgrd}

\item{forumas}{a list of length two specifying the equations for
the center and radius. Changes from the default equation are only relevant
if the user wants to center or scale the intervals according to another
variable(s).}

\item{...}{additional arguments to gstat::gstat().}

\item{x}{an object of class intsp}

\item{forumas}{a list of length two specifying the equations for
the center and radius. Changes from the default equation are only relevant
if the user wants to center or scale the intervals according to another
variable(s).}

\item{...}{additional arguments to gstat::gstat().}
}
\description{
Function to fit empirical variograms for an interval-valued spatial object

Function to create a variogram object for interval-valued data.
A wrapper to gstat::gstat() that fits all of the empirical variograms for
the interval slot of an intgrd object.

Function to create a variogram object for interval-valued data.
A wrapper to gstat::gstat() that fits all of the empirical variograms for
the interval slot of an intsp object.
}
