% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat-extensions.R
\name{intvCheck}
\alias{intvCheck}
\title{Function to visualize the three variograms from an interval valued
spatial data frame.}
\usage{
intvCheck(x, models, ...)
}
\arguments{
\item{x}{an object of class intvariogram}

\item{models}{a list of fitted variogram models, typically an output of
fit.intvariogram}

\item{...}{additional arguments to gstat::plot.variogram()}
}
\value{
nothing (but do plot to the screen)
}
\description{
Function to visualize the three variograms from an interval valued
spatial data frame.
}
