% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohtemp}
\alias{ohtemp}
\title{30 year trimmed mean daily maximum and minimum temperatures for the
Ohio river basin.}
\format{A data frame with 161 rows and 8 variables:

 \describe{
 \item{ID}{The global historical climatological network (GHCN)
 station identifier}
 \item{NAME}{The GHCN station name}
 \item{STATE}{The two-digit designation for the state in which each station resides}
 \item{LATITUDE}{Latitude coordinate position}
 \item{LONGITUDE}{Longitude coordinate position}
 \item{ELEVATION}{Elevation of the measurement location (meters)}
 \item{minm}{The 30 year mean daily minimum temperature (tenths of degrees Celsius)}
 \item{maxm}{The 30 year mean daily maximum temperature (tenths of degrees Celsius)}

 }}
\usage{
ohtemp
}
\description{
Intervals are defined by the mean daily maximum and minimum temperatures for
the Ohio river basin from January 1, 1988 - December 31, 2018. The 116
observations in this dataset all had at least 300 daily observations of
temperature in at least 30 of the 31 considered years. The mean was
calculated after trimming 10% of the data in the tails to remove the
influence of potential outliers.
}
\keyword{datasets}
