% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\docType{methods}
\name{plot,intgrd,missing-method}
\alias{plot,intgrd,missing-method}
\title{Create an interval plot for spatial grid.}
\usage{
\S4method{plot}{intgrd,missing}(x, beside = TRUE, circleCol = "black",
  radPos = 2, cex.rad = 1, ...)
}
\arguments{
\item{x}{an object of class intsp or intgrd}

\item{beside}{if true, center and radius plotted side by side
if false, center and radius are plotted in a single figure with the center
plotted using color and the radius plotted using circles circumsribed
within each grid cell.}

\item{circleCol}{if beside=TRUE, the color of the circles
that will be circumscribed within each grid cell}

\item{radPos}{the position of the reference circle for the radius legend
represented as an integer from 1 to 4. Each value corresponds to a
corner of the plot starting in the lower left corner and proceeding
counter-clockwise.}

\item{cex.rad}{controls the size of the text in the Radius legend.}

\item{...}{additional arguments to sp::spplot()}
}
\description{
Calls the sp::spplot() function to plot the locations, centers, and
radii of an interval-valued spatial data frame in a single figure.
}
