% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{plot}
\alias{plot}
\alias{plot,intgrd-method}
\alias{plot,intsp-method}
\title{Create an interval plot for spatial points.}
\arguments{
\item{x}{an object of class intsp or intgrd}

\item{beside}{(intgrd only) if true, center and radius plotted side by side
if false, center and radius are plotted in a single figure with the center
plotted using color and the radius plotted using circles circumsribed
within each grid cell.}

\item{circleCol}{(intgrdonly) if beside=TRUE, the color of the circles
that will be circumscribed within each grid cell}

\item{locationsOnly}{a boolean: TRUE calls plot.sp(x)}

\item{legend.positions}{the positions of the center and radius legend
relative to the plotting window}

\item{cuts}{the number of ranges of values to print in the center and radius
legend respectively}

\item{radSize}{a vector of length 2 indicating the range of point sizes to
plot to visualize radii magnitudes}

\item{pch}{the shape of the points (see plot())}

\item{alpha}{the transparency of the points}

\item{...}{additional arguments to sp::spplot()}
}
\description{
Calls the sp::spplot() function to plot the locations, centers, and
radii of an interval-valued spatial data frame in a single figure.
}
