% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{plot.intsp}
\alias{plot.intsp}
\title{Create an interval plot for spatial points.}
\usage{
\method{plot}{intsp}(x, locationsOnly = FALSE,
  legend.positions = c("left", "right"), cuts = c(5, 5),
  radSize = c(0.1, 3), pch = 16, alpha = 0.5, ...)
}
\arguments{
\item{x}{an object of class intsp}

\item{locationsOnly}{a boolean: TRUE calls plot.sp(x)}

\item{legend.positions}{the positions of the center and radius legend
relative to the plotting window}

\item{cuts}{the number of ranges of values to print in the center and radius
legend respectively}

\item{radSize}{a vector of length 2 indicating the range of point sizes to
plot to visualize radii magnitudes}

\item{pch}{the shape of the points (see plot())}

\item{alpha}{the transparency of the points}

\item{...}{additional arguments to sp::spplot()}
}
\value{
a lattice plot object
}
\description{
Calls the sp::spplot() function to plot the locations, centers, and
radii of an interval-valued spatial data frame in a single figure.
}
