% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{print.intgrd}
\alias{print.intgrd}
\title{Print the contents of an intgrd object}
\usage{
\method{print}{intgrd}(x, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{an object of class intgrd}

\item{...}{additional arguments to base::print}

\item{digits}{default option from sp package}
}
\value{
nothing
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
