% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{print.intsp}
\alias{print.intsp}
\title{Print the contents of an \code{intsp} object}
\usage{
\method{print}{intsp}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class \code{intsp}.}

\item{digits}{Determines how numbers are displayed on the screen.
Default option taken from \code{sp} package.}

\item{...}{Additional arguments to \code{\link{print}}.}
}
\value{
Prints object to the screen, identical to
  \code{\link[sp]{SpatialPoints-class}}, as well as summary statistics for the
  interval slot.
}
\description{
This function extends printing methods in the \code{sp} package
  by including a display and summary of the interval slot for
  the object.
}
