% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{print.summary.intgrd}
\alias{print.summary.intgrd}
\title{Print the object summary to the screen.}
\usage{
\method{print}{summary.intgrd}(x, ...)
}
\arguments{
\item{x}{an object an object of class intgrd}

\item{...}{additional arguments to print()}
}
\value{
nothing
}
\description{
Print the object summary to the screen.
}
