% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{print.summary.intsp}
\alias{print.summary.intsp}
\title{Print the object summary to the screen.}
\usage{
\method{print}{summary.intsp}(x, ...)
}
\arguments{
\item{x}{An object an object of class \code{intsp}.}

\item{...}{Additional arguments to \code{\link{print}}.}
}
\value{
Prints a series of summaries to the screen, identical to
  \code{\link[sp]{SpatialPoints-class}}, as well as summary statistics for the
  interval slot.
}
\description{
Print the object summary to the screen.
}
