% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{show,intgrd-method}
\alias{show,intgrd-method}
\title{Extension of the show function for intgrd objects}
\usage{
\S4method{show}{intgrd}(object)
}
\arguments{
\item{object}{and object of class intgrd}
}
\description{
Extension of the show function for intgrd objects
}
