% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{show,intsp-method}
\alias{show,intsp-method}
\title{Extension of the show function for intgrd objects}
\usage{
\S4method{show}{intsp}(object)
}
\arguments{
\item{object}{An object of class \code{intsp}.}
}
\description{
Extension of the show function for intgrd objects
}
