% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{summary.intgrd}
\alias{summary.intgrd}
\title{Summarize the contents of an \code{intgrd} object,
including special summaries for the interval slot.}
\usage{
\method{summary}{intgrd}(object, ...)
}
\arguments{
\item{object}{An object of class \code{intgrd}.}

\item{...}{Additional arguments to \code{\link{summary}}.}
}
\value{
Prints a summary of the object observations to the screen,
  identical to \code{\link[sp]{SpatialPoints-class}}, as well as
  summary statistics for the interval slot.
}
\description{
Summarize the contents of an \code{intgrd} object,
including special summaries for the interval slot.
}
