% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{summary.intsp}
\alias{summary.intsp}
\title{Summarize the contents of an \code{intsp} object,
  including special summaries for the interval slot.}
\usage{
\method{summary}{intsp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{intsp}.}

\item{...}{Additional arguments to \code{\link{summary}}.}
}
\value{
Prints a series of summaries to the screen, identical to
  \code{\link[sp]{SpatialPoints-class}}, as well as summary statistics for the
  interval slot.
}
\description{
Summarize the contents of an \code{intsp} object,
  including special summaries for the interval slot.
}
