% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{tail.intgrd}
\alias{tail.intgrd}
\title{Print the tail of an intgrd object.}
\usage{
\method{tail}{intgrd}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{an object of class intgrd}

\item{n}{number of rows to print to the screen}

\item{...}{additional arguments to utils::tail()}

\item{digits}{determines how numbers are displayed (default taken from
package sp)}
}
\value{
nothing
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
