% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{extract}
\alias{extract}
\alias{$}
\alias{$,intgrd-method}
\alias{$<-}
\alias{$<-,intgrd-method}
\alias{[,intgrd,ANY,missing,missing-method}
\alias{[,intgrd-method}
\alias{$,intsp-method}
\alias{$<-,intsp-method}
\alias{[,intsp,ANY,missing,missing-method}
\alias{[,intsp-method}
\title{Extract subset of an \code{intsp} or \code{intgrd} object}
\usage{
\S4method{$}{intgrd}(x, name)

\S4method{$}{intgrd}(x, name) <- value

\S4method{[}{intgrd,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{$}{intsp}(x, name)

\S4method{$}{intsp}(x, name) <- value

\S4method{[}{intsp,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{$}{intsp}(x, name) <- value
}
\arguments{
\item{x}{An object of class \code{intsp} or class \code{intgrd}
from which to replace elements.}

\item{name}{Character vector corresponding to the name of the column that
will be extracted or replaced.}

\item{value}{The new data used to replaced drop data in the desired slot.}

\item{i, j, ...}{indices specifying elements to extract or replace. See
generic function documentation for details.}

\item{drop}{The requested column that may be reassigned.}
}
\value{
An object of class \code{intsp} or \code{intgrd}.
}
\description{
Extract subset of an \code{intsp} or \code{intgrd} object
}
