% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{$}
\alias{$}
\alias{$,intgrd-method}
\alias{$,intsp-method}
\title{Atomic subsetting of interval spatial objects.}
\arguments{
\item{x}{an object of class \'intsp\' or class \'intgrd\'}

\item{name}{the variable name on which to subset. This variable name can be
located in the data, coordinates, or interval slot.}
}
\value{
the value contained in the requested subset
}
\description{
Atomic subsetting of interval spatial objects.
}
