% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat-extensions.R
\name{fit.intvariogram}
\alias{fit.intvariogram}
\title{Function to create a variogram object for interval-valued data.}
\usage{
fit.intvariogram(x, models = gstat::vgm(rep("Sph", 3)), ...)
}
\arguments{
\item{x}{an object of class intvariogram}

\item{models}{an object of class variogramModelList. Must specify at least
two variogram models to fit (for center and radius) if less than three
models are specified then the method fails to fit a variogram for the
center radius interaction.}

\item{...}{additional arguments to gstat::fit.variogram()}
}
\value{
a list containing the parameters for each of the three (or two)
fitted variograms.
}
\description{
Function to create a variogram object for interval-valued data.
}
