% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{head.intgrd}
\alias{head.intgrd}
\title{Print the head of an intgrd object.}
\usage{
\method{head}{intgrd}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{an object of class intgrd}

\item{n}{number of rows to print to the screen}

\item{...}{additional arguments to print}

\item{digits}{determines how values are printed to the screen
(default from sp package)}
}
\value{
nothing
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
