% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{head.intsp}
\alias{head.intsp}
\title{Print the head of an intsp object.}
\usage{
\method{head}{intsp}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{an object of class intsp}

\item{n}{number of rows to print to the screen}

\item{...}{additional arguments to print}

\item{digits}{determines how values are printed to the screen
(default from sp package)}
}
\value{
nothing
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
