% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,intgrd-method}
\alias{as.data.frame,intsp-method}
\title{Convert intgrd or intsp object back to a data frame}
\usage{
\S4method{as.data.frame}{intgrd}(x)

\S4method{as.data.frame}{intsp}(x)
}
\arguments{
\item{x}{An object of class \code{intsp} or class \code{intgrd}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Convert intgrd or intsp object back to a data frame
}
