% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{intkrige-package}
\alias{intkrige-package}
\alias{_PACKAGE}
\title{intkrige package: Interval-valued kriging models in R.}
\description{
An interval valued extension of ordinary and simple kriging. 
    Optimization of the function is based on a generalized interval distance. 
    This creates a non-differentiable cost function that requires a 
    differentiable approximation to the absolute value function. This 
    differentiable appoximation is optimized using a Newton-Raphson algorithm
    with a penalty function to impose the constraints. Covariances are 
    specified using the vgm() object conventions in the gstat package. 
    These functions use the gstat variogram conventions.
}
\section{Description}{

An interval valued extension of ordinary and simple kriging.
Optimization of the function is based on a generalized interval distance.
This creates a non-differentiable cost function that requires a
differentiable approximation to the absolute value function. This
differentiable appoximation is optimized using a Newton-Raphson algorithm
with a penalty function to impose the constraints.
Covariances are specified using the \code{\link[gstat]{vgm}}
object conventions.
}

\section{Functions}{

\itemize{
\item \code{\link{intkrige}} Predict using an interval-valued kriging model.
\item \code{\link{dist_cpp}} Function to compute geographic or euclidean
distances.
}
}

\section{Data}{

\itemize{
\item \link{UTsnow} An interval-valued design ground snow load dataset for Utah.
}
}

\author{
Brennan Bean \email{brennan.bean.20@gmail.com}
}
