% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/intgrd-class.R, R/intsp-class.R
\name{intvariogram}
\alias{intvariogram}
\alias{intvariogram,intgrd-method}
\alias{intvariogram,intsp-method}
\title{Function to fit empirical variograms for an interval-valued spatial object}
\usage{
intvariogram(x, formulas = list(center ~ 1, radius ~ 1), ...)

\S4method{intvariogram}{intgrd}(x, formulas = list(center ~ 1, radius ~ 1), ...)

\S4method{intvariogram}{intsp}(x, formulas = list(center ~ 1, radius ~ 1), ...)
}
\arguments{
\item{x}{An object of class intsp or intgrd.}

\item{formulas}{A list of length two specifying the formulas related to the
centers and radii respectively.}

\item{...}{Additional arguments for sp::variogram().}
}
\value{
An object of class 'intvariogram' containing empirical variograms
  for the center, radius, and center/radius interaction.
}
\description{
Function to fit empirical variograms for an interval-valued spatial object
}
