% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstat-extensions.R
\name{intvCheck}
\alias{intvCheck}
\title{Function to visualize the three variograms from an interval valued
spatial data frame.}
\usage{
intvCheck(x, models, ...)
}
\arguments{
\item{x}{An object of class \code{intvariogram}.}

\item{models}{A list of fitted variogram models, typically an output of
\code{\link{fit.intvariogram}}.}

\item{...}{Additional arguments to
\code{\link[gstat]{plot.gstatVariogram}}.}
}
\description{
Function to visualize the three variograms from an interval valued
spatial data frame.
}
