% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohMap}
\alias{ohMap}
\title{Ohio River Basin map}
\format{A SpatialPolygons objects with the boundaries of
 the Ohio river basin.}
\usage{
ohMap
}
\description{
Ohio River Basin map
}
\keyword{datasets}
