% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\docType{methods}
\name{plot,intgrd,missing-method}
\alias{plot,intgrd,missing-method}
\title{Create an interval plot for spatial grid.}
\usage{
\S4method{plot}{intgrd,missing}(x, beside = TRUE, circleCol = "black",
  minRad = 0.25, ...)
}
\arguments{
\item{x}{an object of class intgrd}

\item{beside}{if true, center and radius plotted side by side
if false, center and radius are plotted in a single figure with the center
plotted using color and the radius plotted using circles circumsribed
within each grid cell.}

\item{circleCol}{if beside=TRUE, the color of the circles
that will be circumscribed within each grid cell}

\item{minRad}{the value of the minimum value of the radius in the circles
drawn to represent the interval radii. Must be a number between 0 and 1.}

\item{...}{additional arguments to sp::spplot()}
}
\description{
Calls the sp::spplot() function to plot the locations, centers, and
radii of an interval-valued spatial data frame in a single figure.
}
