% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{print.intgrd}
\alias{print.intgrd}
\title{Print the contents of an intgrd object}
\usage{
\method{print}{intgrd}(x, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{An object of class \code{intgrd}.}

\item{...}{Additional arguments to \code{\link{print}}.}

\item{digits}{Determines how numeric values are printed to the screen
(default from \code{sp} package).}
}
\value{
Prints object to the screen, identical to
  \code{\link[sp]{SpatialPoints-class}}, as well as summary statistics for the
  interval slot.
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
