% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{print.intsp}
\alias{print.intsp}
\title{Print the contents of an intsp object}
\usage{
\method{print}{intsp}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{an object of class intsp}

\item{digits}{default option from sp package}

\item{...}{additional arguments to base::print}
}
\value{
nothing
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
