% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{print.summary.intgrd}
\alias{print.summary.intgrd}
\title{Print the object summary to the screen.}
\usage{
\method{print}{summary.intgrd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{intgrd}.}

\item{...}{Additional arguments to \code{\link{print}}.}
}
\value{
Prints a subset of the object observations to the screen,
  identical to \code{\link[sp]{SpatialPoints-class}}, as well as
  summary statistics for the interval slot.
}
\description{
Print the object summary to the screen.
}
