% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{print.summary.intsp}
\alias{print.summary.intsp}
\title{Print the object summary to the screen.}
\usage{
\method{print}{summary.intsp}(x, ...)
}
\arguments{
\item{x}{an object an object of class intsp}

\item{...}{additional arguments to base::print()}
}
\value{
nothing
}
\description{
Print the object summary to the screen.
}
