% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{summary.intgrd}
\alias{summary.intgrd}
\title{Summarize the contents of an intgrd object, including special summaries for
the interval slot.}
\usage{
\method{summary}{intgrd}(object, ...)
}
\arguments{
\item{object}{an object of class intgrd}

\item{...}{additional arguments to base::summary().}
}
\value{
nothing
}
\description{
Summarize the contents of an intgrd object, including special summaries for
the interval slot.
}
