% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{summary.intsp}
\alias{summary.intsp}
\title{Summarize the contents of an intsp object, including special summaries for
the interval slot.}
\usage{
\method{summary}{intsp}(object, ...)
}
\arguments{
\item{object}{an object of class intsp}

\item{...}{additional arguments to base::summary().}
}
\value{
nothing
}
\description{
Summarize the contents of an intsp object, including special summaries for
the interval slot.
}
