% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intgrd-class.R
\name{tail.intgrd}
\alias{tail.intgrd}
\title{Print the tail of an intgrd object.}
\usage{
\method{tail}{intgrd}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{An object of class \code{intgrd}.}

\item{n}{The number of rows to print to the screen.}

\item{...}{Additional arguments to \code{tail}.}

\item{digits}{Determines how numbers are displayed to the screen
(default taken from package \code{sp}).}
}
\value{
Prints a subset of the object observations to the screen,
  identical to \code{\link[sp]{SpatialPoints-class}}, as well as
  summary statistics for the interval slot.
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
