% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsp-class.R
\name{tail.intsp}
\alias{tail.intsp}
\title{Print the tail of an intsp object.}
\usage{
\method{tail}{intsp}(x, n = 6, ..., digits = getOption("digits"))
}
\arguments{
\item{x}{An object of class \code{intsp}.}

\item{n}{The number of rows to print to the screen.}

\item{...}{Additional arguments to \code{\link{tail}}.}

\item{digits}{Determines how numbers are displayed (default taken from
sp package).}
}
\value{
Prints a subset of the object to the screen, identical to
  \code{\link[sp]{SpatialPoints-class}}, as well as summary statistics for the
  interval slot.
}
\description{
This function extends print.sp by including a display and summary of the
interval slot for the object.
}
